+++ dart/common/free_list_allocator.hpp
@@ -40,6 +40,8 @@
 
 #include <vector>
 
+#include <cstddef>
+
 namespace dart::common {
 
 /// Most general heap memory allocator for allocating memory of various sizes.
@@ -100,7 +102,7 @@ public:
   void print(std::ostream& os = std::cout, int indent = 0) const override;
 
 private:
-  struct MemoryBlockHeader
+  struct alignas(std::max_align_t) MemoryBlockHeader
   {
     /// Memory block size in bytes
     size_t mSize;